Segment: Basic accessors and loaders
Segment - A structure that holds multiple segment_column(s) together.

Currently loading is done through - supplying a data directory and all
the files with the extension '.mastidb' will be considered as part of
the segment. These are segment column data files.
The name of the file becomes the column name as a convention.
Metadata files for all columns are preloaded and deserialized during segment load.
Segment column data is read/deserialized as it is required.